loader/requirements.txt
psycopg2-binary==2.7.4
pika==0.11.2